README.m
# 2048

Greetings to whoever has stumbled upon this project and is reading this readme. This is my first self-driven project. Its aim is to be a fully fledged 2048 browser game. The concept of the game is entirely borrowed from the popular 2048 game. 
The game logic is built using JavaScript and basic DOM manipulations, without additional frameworks/libraries. The stylesheets are rendered using SCSS. The basic HTML file only puts the skeleton of the page. The game grid is rendered within the JS logic. It supports good size responsiveness, where it would decrease the game grid and font size on smaller screens. Game interaction is performed using arrow keys for regular users, as well as swipe direction detection for touchscreens(tablets, smartphones etc.). For movement animation effect, the app uses a second grid, which comes visible only when movement is performed and only for squares that are moved. The movement is performed using simple CSS translate and transition rules and no "animation" features. There are additional visual effects for when numbers collide and add up and to display score increasement. Grid layout system is used to display the game grid. The SCSS utilizes a loop to produce class for each separate number tag, generating a distinguishable HSL background color and fluid font color, as well as corresponding font size to fit larger numbers within their element containment. 
There is a game-over effect as well as "new game" button to initiate a new game.